requirements.txt

Flask==2.2.2
google-cloud-storage==2.7.0
google-generativeai==0.4.0
gunicorn==20.1.0
bcrypt==4.0.1
